transactions

+++ controller/expenseController.js
@@ -1,17 +1,22 @@
 const { Users } = require("../models");
 const Expense = require("../models/exepnse");
+const sequelize = require("../utils/db-conection");
+ 
 
 const addExpense = async(req,res)=>{
     try{
+        const transaction = await sequelize.transaction();
         const expenses = {...req.body,UserId:req.user.id}
-        const expense = await Expense.create(expenses);
+        const expense = await Expense.create(expenses,{transaction});
         if(expense){
             const user = await Users.findByPk(req.user.id);
             user.totalExpense = user.totalExpense+Number(req.body.amount);
-            await user.save();
+            await user.save({transaction});
+            await transaction.commit();
             return res.status(201).json({expense});
         }
     }catch(err){
+        await transaction.rollback();
         return res.status(500).json({message:"Something went wrong"});
     }
 }
@@ -32,25 +37,49 @@ const getAllExpenses = async(req,res)=>{
 
 }
 
-const deleteExpense = async(req,res)=>{
-    try{
-        const expenseId = req.params.id;
-        const deleted = await Expense.destroy({
-            where:{
-                id:expenseId,
-                UserId:req.user.id
-            }
-        })
-        if (!deleted) {
-            return res.status(404).json({ message: "Expense not found" });
-          }
-      
-          return res.status(200).json({ message: "Expense deleted successfully" });
-      
-    }catch(err){
-        return res.status(500).json({message:"Something Went Wrong"});
-    }
+const deleteExpense = async (req, res) => {
+    let transaction;
 
-}
+    try {
+        transaction = await sequelize.transaction();
+
+        const expenseId = req.params.id;
+
+        const expense = await Expense.findOne({
+            where: {
+                id: expenseId,
+                UserId: req.user.id
+            },
+            transaction
+        });
+
+        if (!expense) {
+            await transaction.rollback();
+            return res.status(404).json({ message: "Expense not found" });
+        }
+
+        const user = await Users.findByPk(req.user.id, { transaction });
+
+        user.totalExpense -= expense.amount;
+        await user.save({ transaction });
+
+        await Expense.destroy({
+            where: {
+                id: expenseId,
+                UserId: req.user.id
+            },
+            transaction
+        });
+
+        await transaction.commit();
+
+        return res.status(200).json({ message: "Expense deleted successfully" });
+
+    } catch (err) {
+        if (transaction) await transaction.rollback();
+        console.log(err);
+        return res.status(500).json({ message: "Something Went Wrong" });
+    }
+};
 
 module.exports = {addExpense,getAllExpenses,deleteExpense}
